Refactor egm modules
