+++ src/dcegm/egm/solve_euler_equation.py
